First draft removing ip check feature and adding ipcheck for validator
