+++ full-service/src/config.rs
@@ -19,11 +19,7 @@ use mc_util_uri::{ConnectionUri, ConsensusClientUri, FogUri};
 use mc_validator_api::ValidatorUri;
 
 use displaydoc::Display;
-#[cfg(feature = "ip-check")]
-use reqwest::{
-    blocking::Client,
-    header::{HeaderMap, HeaderValue, CONTENT_TYPE},
-};
+
 use std::{
     convert::TryFrom,
     fs,
@@ -33,32 +29,7 @@ use std::{
 };
 use structopt::StructOpt;
 
-#[derive(Display, Debug)]
-pub enum ConfigError {
-    /// Error parsing json {0}
-    Json(serde_json::Error),
-
-    /// Error handling reqwest {0}
-    Reqwest(reqwest::Error),
-
-    /// Invalid country
-    InvalidCountry,
-
-    /// Data missing in the response {0}
-    DataMissing(String),
-}
-
-impl From<serde_json::Error> for ConfigError {
-    fn from(e: serde_json::Error) -> Self {
-        Self::Json(e)
-    }
-}
-
-impl From<reqwest::Error> for ConfigError {
-    fn from(e: reqwest::Error) -> Self {
-        Self::Reqwest(e)
-    }
-}
+use crate::check_host::{self, CheckHostError};
 
 /// Command line config for the Wallet API
 #[derive(Clone, Debug, StructOpt)]
@@ -189,43 +160,11 @@ impl APIConfig {
     /// Uses ipinfo.io for getting details about IP address.
     ///
     /// Note, both of these services are free tier and rate-limited.
-    #[cfg(feature = "ip-check")]
-    pub fn validate_host(&self) -> Result<(), ConfigError> {
-        let client = Client::builder().gzip(true).use_rustls_tls().build()?;
-        let mut json_headers = HeaderMap::new();
-        json_headers.insert(CONTENT_TYPE, HeaderValue::from_static("application/json"));
-        let response = client
-            .get("https://ipinfo.io/json/")
-            .headers(json_headers)
-            .send()?
-            .error_for_status()?;
-        let data = response.text()?;
-        let data_json: serde_json::Value = serde_json::from_str(&data)?;
-
-        let data_missing_err = Err(ConfigError::DataMissing(data_json.to_string()));
-        let country: &str = match data_json["country"].as_str() {
-            Some(c) => c,
-            None => return data_missing_err,
-        };
-        let region: &str = match data_json["region"].as_str() {
-            Some(r) => r,
-            None => return data_missing_err,
-        };
-
-        let err = Err(ConfigError::InvalidCountry);
-        match country {
-            "IR" | "SY" | "CU" | "KP" => err,
-            "UA" => match region {
-                "Crimea" => err,
-                _ => Ok(()),
-            },
-            _ => Ok(()),
+    pub fn validate_host(&self) -> Result<(), CheckHostError> {
+        if self.validator.is_some() {
+            return Ok(());
         }
-    }
-
-    #[cfg(not(feature = "ip-check"))]
-    pub fn validate_host(&self) -> Result<(), ConfigError> {
-        Ok(())
+        check_host::check_host_is_allowed_country_and_region()
     }
 }
