+++ full-service/src/lib.rs
@@ -4,6 +4,7 @@
 
 #![feature(proc_macro_hygiene, decl_macro)]
 
+pub mod check_host;
 pub mod config;
 mod db;
 mod error;
